+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/PresentationFactory.kt
@@ -20,6 +20,7 @@ import at.asitplus.signum.indispensable.josef.toJsonWebKey
 import at.asitplus.wallet.lib.agent.*
 import at.asitplus.wallet.lib.cbor.CoseService
 import at.asitplus.wallet.lib.data.CredentialPresentation
+import at.asitplus.wallet.lib.data.DeprecatedBase64URLTransactionDataSerializer
 import at.asitplus.wallet.lib.data.dif.PresentationSubmissionValidator
 import at.asitplus.wallet.lib.data.vckJsonSerializer
 import at.asitplus.wallet.lib.iso.*
@@ -34,7 +35,7 @@ import kotlinx.datetime.Clock
 import kotlinx.serialization.PolymorphicSerializer
 import kotlinx.serialization.builtins.ByteArraySerializer
 import kotlinx.serialization.encodeToByteArray
-import kotlinx.serialization.json.decodeFromJsonElement
+import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.jsonArray
 import kotlin.coroutines.cancellation.CancellationException
 import kotlin.random.Random
@@ -90,17 +91,23 @@ internal class PresentationFactory(
     /**
      * Parses all `transaction_data` fields from the request, with a JsonPath, because
      * ... for OpenID4VP Draft 23, that's encoded in the AuthnRequest
-     * ... but for Potential UC 5, that's encoded in the input descriptor,
-     *     and we don't have access to that class from the module vck-rqes
+     * ... but for Potential UC 5, that's encoded in the input descriptor
+     *     and we cannot deserialize into data classes defined in [at.asitplus.rqes]
      */
-    private fun parseTransactionData(request: RequestParameters): Set<ByteArray>? =
+    private fun parseTransactionData(request: RequestParameters): Collection<TransactionData>? =
         with(vckJsonSerializer.encodeToJsonElement(PolymorphicSerializer(RequestParameters::class), request)) {
             JsonPath("$..transaction_data").query(this)
                 .flatMap { it.value.jsonArray }
-                .map { vckJsonSerializer.decodeFromJsonElement<String>(it) }
-                .filter { it.isNotEmpty() }
-                .map { if (it.contains(",")) it.encodeToByteArray() else it.decodeToByteArray(Base64UrlStrict) }
-                .toSet()
+                .map { vckJsonSerializer.encodeToString<JsonElement>(it) }
+                .mapNotNull {
+                    runCatching {
+                        vckJsonSerializer.decodeFromString(
+                            DeprecatedBase64URLTransactionDataSerializer,
+                            it
+                        )
+                    }.getOrNull()
+                }
+                .distinct()
                 .ifEmpty { null }
         }