+++ vck-rqes/src/commonMain/kotlin/at/asitplus/wallet/lib/rqes/RqesOpenId4VpVerifier.kt
@@ -6,11 +6,9 @@ import at.asitplus.dif.InputDescriptor
 import at.asitplus.dif.PresentationDefinition
 import at.asitplus.openid.AuthenticationRequestParameters
 import at.asitplus.rqes.QesInputDescriptor
-import at.asitplus.rqes.collection_entries.TransactionData
 import at.asitplus.wallet.lib.agent.*
 import at.asitplus.wallet.lib.cbor.DefaultVerifierCoseService
 import at.asitplus.wallet.lib.cbor.VerifierCoseService
-import at.asitplus.wallet.lib.data.vckJsonSerializer
 import at.asitplus.wallet.lib.jws.DefaultJwsService
 import at.asitplus.wallet.lib.jws.DefaultVerifierJwsService
 import at.asitplus.wallet.lib.jws.JwsService
@@ -23,14 +21,14 @@ import at.asitplus.wallet.lib.openid.ClientIdScheme
 import at.asitplus.wallet.lib.openid.OpenId4VpVerifier
 import at.asitplus.wallet.lib.openid.OpenIdRequestOptions
 import at.asitplus.wallet.lib.openid.RequestOptions
-import at.asitplus.wallet.lib.rqes.helper.OpenIdRqesParameters
 import com.benasher44.uuid.uuid4
+import io.ktor.util.*
 import kotlinx.datetime.Clock
-import kotlinx.serialization.encodeToString
 
 /**
  * Verifier with access to [TransactionData] class can now generate requests containing [TransactionData]
  */
+@Deprecated("OpenId4VpVerifier can now access TransactionData, for RqesRequests use RqesRequestOptions", ReplaceWith("OpenId4VpVerifier"))
 class RqesOpenId4VpVerifier(
     private val clientIdScheme: ClientIdScheme,
     private val keyMaterial: KeyMaterial = EphemeralKeyWithoutCert(),
@@ -56,11 +54,12 @@ class RqesOpenId4VpVerifier(
     stateToAuthnRequestStore
 ) {
     /**
-     * ExtendedRequestOptions cannot generate DifInputDescriptors!
+     * Necessary to use [QesInputDescriptor]
+     * ExtendedRequestOptions cannot generate [DifInputDescriptor]!
      */
+    @Deprecated("Replaced", ReplaceWith("RqesRequestOptions"))
     data class ExtendedRequestOptions(
         val baseRequestOptions: OpenIdRequestOptions,
-        val rqesParameters: OpenIdRqesParameters,
     ) : RequestOptions by baseRequestOptions {
 
         override fun toPresentationDefinition(
@@ -79,25 +78,8 @@ class RqesOpenId4VpVerifier(
                 id = requestOptionCredential.buildId(),
                 format = requestOptionCredential.toFormatHolder(containerJwt, containerSdJwt),
                 constraints = requestOptionCredential.toConstraint(),
-                transactionData = rqesParameters.transactionData.toList()
+                transactionData = transactionData?.toList()
             )
         }
     }
-
-    override suspend fun enrichAuthnRequest(
-        params: AuthenticationRequestParameters,
-        requestOptions: RequestOptions,
-    ): AuthenticationRequestParameters = with(requestOptions) {
-        when (this) {
-            is OpenIdRequestOptions -> params
-            is ExtendedRequestOptions -> params.copy(
-                transactionData = this.rqesParameters.transactionData.map {
-                    vckJsonSerializer.encodeToString(it)
-                }.toSet()
-            )
-
-            else -> throw NotImplementedError("Unknown RequestOption class: ${this::class}")
-        }
-    }
-
 }