+++ vck-rqes/src/commonTest/kotlin/at/asitplus/wallet/lib/rqes/RqesRequestOptionsTest.kt
@@ -1,30 +1,19 @@
 package at.asitplus.wallet.lib.rqes
 
 import at.asitplus.openid.SignatureQualifier
+import at.asitplus.openid.TransactionData
 import at.asitplus.rqes.QesInputDescriptor
+import at.asitplus.rqes.collection_entries.QesAuthorization
 import at.asitplus.rqes.collection_entries.RqesDocumentDigestEntry
-import at.asitplus.rqes.collection_entries.TransactionData
 import at.asitplus.signum.indispensable.Digest
 import at.asitplus.wallet.eupid.EuPidScheme
 import at.asitplus.wallet.eupid.EuPidScheme.SdJwtAttributes.FAMILY_NAME
 import at.asitplus.wallet.eupid.EuPidScheme.SdJwtAttributes.GIVEN_NAME
-import at.asitplus.wallet.lib.agent.EphemeralKeyWithoutCert
-import at.asitplus.wallet.lib.agent.Holder
-import at.asitplus.wallet.lib.agent.HolderAgent
-import at.asitplus.wallet.lib.agent.IssuerAgent
-import at.asitplus.wallet.lib.agent.KeyMaterial
-import at.asitplus.wallet.lib.agent.toStoreCredentialInput
+import at.asitplus.wallet.lib.agent.*
 import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation.SD_JWT
-import at.asitplus.wallet.lib.data.vckJsonSerializer
 import at.asitplus.wallet.lib.oidvci.encodeToParameters
-import at.asitplus.wallet.lib.openid.AuthenticationResponseResult
-import at.asitplus.wallet.lib.openid.AuthnResponseResult
-import at.asitplus.wallet.lib.openid.ClientIdScheme
-import at.asitplus.wallet.lib.openid.OpenId4VpHolder
+import at.asitplus.wallet.lib.openid.*
 import at.asitplus.wallet.lib.openid.OpenId4VpVerifier.CreationOptions.Query
-import at.asitplus.wallet.lib.openid.OpenIdRequestOptions
-import at.asitplus.wallet.lib.openid.RequestOptionsCredential
-import at.asitplus.wallet.lib.rqes.helper.OpenIdRqesParameters
 import com.benasher44.uuid.bytes
 import com.benasher44.uuid.uuid4
 import io.kotest.core.spec.style.FreeSpec
@@ -34,10 +23,9 @@ import io.kotest.matchers.shouldNotBe
 import io.kotest.matchers.string.shouldContain
 import io.kotest.matchers.types.shouldBeInstanceOf
 import io.ktor.http.*
-import kotlinx.serialization.encodeToString
 
 @Suppress("DEPRECATION")
-class RqesOpenId4VpVerifierTest : FreeSpec({
+class RqesRequestOptionsTest : FreeSpec({
 
     lateinit var clientId: String
     lateinit var walletUrl: String
@@ -45,8 +33,8 @@ class RqesOpenId4VpVerifierTest : FreeSpec({
     lateinit var verifierKeyMaterial: KeyMaterial
     lateinit var holderAgent: Holder
     lateinit var holderOid4vp: OpenId4VpHolder
-    lateinit var rqesVerifier: RqesOpenId4VpVerifier
-    lateinit var requestOptions: RqesOpenId4VpVerifier.ExtendedRequestOptions
+    lateinit var rqesVerifier: OpenId4VpVerifier
+    lateinit var requestOptions: RqesRequestOptions
 
     beforeEach {
         holderKeyMaterial = EphemeralKeyWithoutCert()
@@ -65,7 +53,7 @@ class RqesOpenId4VpVerifierTest : FreeSpec({
         holderOid4vp = OpenId4VpHolder(
             holder = holderAgent,
         )
-        rqesVerifier = RqesOpenId4VpVerifier(
+        rqesVerifier = OpenId4VpVerifier(
             keyMaterial = verifierKeyMaterial,
             clientIdScheme = ClientIdScheme.RedirectUri(clientId),
         )
@@ -76,7 +64,6 @@ class RqesOpenId4VpVerifierTest : FreeSpec({
 
         "Authentication request contains transaction data" {
             val authnRequest = rqesVerifier.createAuthnRequest(requestOptions = requestOptions)
-
             authnRequest.transactionData shouldNotBe null
             authnRequest.presentationDefinition.shouldNotBeNull()
             authnRequest.presentationDefinition!!.inputDescriptors.first()
@@ -85,10 +72,6 @@ class RqesOpenId4VpVerifierTest : FreeSpec({
         }
 
         "KB-JWT contains transaction data" {
-            val transactionDataEncoded =
-                vckJsonSerializer.encodeToString(requestOptions.rqesParameters.transactionData.first())
-                    .encodeToByteArray()
-
             val authnRequestUrl = rqesVerifier.createAuthnRequest(requestOptions, Query(walletUrl))
                 .getOrThrow().url
             authnRequestUrl shouldContain "transaction_data"
@@ -100,17 +83,13 @@ class RqesOpenId4VpVerifierTest : FreeSpec({
                 .shouldBeInstanceOf<AuthnResponseResult.SuccessSdJwt>()
 
             result.sdJwtSigned.keyBindingJws.shouldNotBeNull().payload.apply {
-                transactionData.shouldNotBeNull().first() shouldBe transactionDataEncoded
+                transactionData.shouldNotBeNull().first() shouldBe requestOptions.transactionData!!.first()
                 transactionDataHashes.shouldNotBeNull()
                 transactionDataHashesAlgorithm.shouldNotBeNull()
             }
         }
 
         "UC5-Specific Flow: KeyBindingJws contains transaction data" {
-            val transactionDataEncoded =
-                vckJsonSerializer.encodeToString(requestOptions.rqesParameters.transactionData.first())
-                    .encodeToByteArray()
-
             //Do not use [AuthenticationRequestParameters.transactionData] introduced in OpenID4VP
             val authnRequest = rqesVerifier.createAuthnRequest(requestOptions).copy(transactionData = null)
 
@@ -126,7 +105,7 @@ class RqesOpenId4VpVerifierTest : FreeSpec({
                 .shouldBeInstanceOf<AuthnResponseResult.SuccessSdJwt>()
 
             result.sdJwtSigned.keyBindingJws.shouldNotBeNull().payload.apply {
-                transactionData.shouldNotBeNull().first() shouldBe transactionDataEncoded
+                transactionData.shouldNotBeNull().first() shouldBe requestOptions.transactionData!!.first()
                 //TODO maybe introduce strict separation of the two flows
                 transactionDataHashes.shouldNotBeNull()
                 transactionDataHashesAlgorithm.shouldNotBeNull()
@@ -135,22 +114,20 @@ class RqesOpenId4VpVerifierTest : FreeSpec({
     }
 })
 
-private fun buildExtendedRequestOptions(): RqesOpenId4VpVerifier.ExtendedRequestOptions =
-    RqesOpenId4VpVerifier.ExtendedRequestOptions(
+private fun buildExtendedRequestOptions(): RqesRequestOptions =
+    RqesRequestOptions(
         baseRequestOptions = OpenIdRequestOptions(
             credentials = setOf(
                 RequestOptionsCredential(
                     EuPidScheme, SD_JWT, setOf(FAMILY_NAME, GIVEN_NAME)
                 )
-            )
-        ),
-        rqesParameters = OpenIdRqesParameters(
+            ),
             transactionData = setOf(getTransactionData())
         )
     )
 
 //TODO other transactionData
-private fun getTransactionData(): TransactionData = TransactionData.QesAuthorization.create(
+private fun getTransactionData(): TransactionData = QesAuthorization.create(
     documentDigest = listOf(getDocumentDigests()),
     signatureQualifier = SignatureQualifier.EU_EIDAS_QES,
     credentialId = uuid4().toString(),